host: freebsd Rlimit uint64 fix + ze-platform-vet gate (AC-7)
spec-followup-subsystem AC-7. platform_other.go assigned
syscall.Rlimit.Cur/.Max (int64 on freebsd, uint64 on linux/darwin) to
uint64 PlatformInfo fields, failing GOOS=freebsd vet. Wrap in uint64()
(no-op on darwin). Add a ze-platform-vet make gate running GOOS=darwin +
GOOS=freebsd go vet over host + iface + plugins/iface, wired into
_ze-verify-impl, _ze-verify-changed-impl, and
scripts/status/verify_run.go stagesForMode (both branches, classified as
a vet stage). Existing _other.go stubs already give darwin/freebsd
interface parity, so no new _darwin.go/_bsd.go files are needed.

+++ internal/component/host/platform_other.go
@@ -20,8 +20,11 @@ func (d *Detector) DetectPlatform() (*PlatformInfo, error) {
 	info.RebootAllowed = false
 	var limit syscall.Rlimit
 	if err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, &limit); err == nil {
-		info.FDLimitSoftCurrent = limit.Cur
-		info.FDLimitHardMax = limit.Max
+		// syscall.Rlimit.Cur/.Max are uint64 on linux/darwin but int64 on
+		// freebsd; the explicit conversions are no-ops on the former and make
+		// this file vet/compile clean under GOOS=freebsd (AC-7).
+		info.FDLimitSoftCurrent = uint64(limit.Cur)
+		info.FDLimitHardMax = uint64(limit.Max)
 		info.FDLimitRaisable = limit.Cur < limit.Max
 	}
 	return info, nil
